Bot/requirements.txt
python-telegram-bot==20.7
requests==2.31.0
urllib3==2.1.0
tldextract==5.1.1
python-dotenv==1.0.0
flask==3.0.0
